Fixing test instrumentation in DataManagerTest.
Issue was caused by the getView() of the QuestinListAdapter returning null
when somewhat blank items are passed to it.

The solution is essentially that we should always check `convertView` and
inflate it if it's null first off, so that we never return null, otherwise
we get some pretty un-debuggable errors.